// apps/documentos/src/types/roles.ts
export enum UserRole {
  SUPERADMIN = 'SUPERADMIN',
  ADMIN = 'ADMIN',
  OPERATOR = 'OPERATOR',
  OPERADOR_INTERNO = 'OPERADOR_INTERNO',
  ADMIN_INTERNO = 'ADMIN_INTERNO',
  DADOR_DE_CARGA = 'DADOR_DE_CARGA',
  TRANSPORTISTA = 'TRANSPORTISTA',
  EMPRESA_TRANSPORTISTA = 'EMPRESA_TRANSPORTISTA',
  CHOFER = 'CHOFER',
  CLIENTE = 'CLIENTE',
}